+++ core/modules/modexternalaccess.class.php
@@ -64,7 +64,7 @@ class modexternalaccess extends DolibarrModules
 
 
 
-		$this->version = '1.21.1';
+		$this->version = '1.21.0';
 
 
 		// Key used in llx_const table to save module status enabled/disabled (where MYMODULE is value of property name of module in uppercase)