+++ ChangeLog.md
@@ -8,6 +8,10 @@ TODO : pour les options commencer à voir si possible de crééer des tab par el
 
 ## Unreleased
 
+
+
+# RELEASE 1.22 - 17/09/2021
+
 - NEW : Gestion des Extrafields de propal sur la liste *14/09/2021* - 1.22.0
 - NEW : Gestion des Extrafields d'expédition sur la liste des BL *19/08/2021* - 1.21.0